% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cellDIVER.R
\name{run_cellDIVER}
\alias{run_cellDIVER}
\title{cellDIVER app}
\usage{
# Option 1: Single-object deployment with
# an object and an object config file
run_cellDIVER(
 object = "path_to_object",
 config_file = "path_to_config_file.yaml"
 )

# Option 2: Multi-object deployment with
# Browser config file with paths to object,
# config files for any number of objects
#
run_cellDIVER(
 browser_config = "path_to_browser_config_file.yaml"
 )
}
\arguments{
\item{browser_config}{path to a YAML config file giving browser specific settings. For more info on generating this config file, see \href{https://amc-heme.github.io/cellDIVER/articles/dataset_setup_walkthrough.html#create-an-app-config-file}{here}.}

\item{object_path}{path to a single-cell object to be configured. Currently,
Seurat, SingleCellExperiment, and anndata objects are supported. For
SingleCellExperiment objects with using HDF5 disk-backed storage via
\link{HDF5Array}, \code{object_path} should be a path to the directory containing
the se.rds and assays.h5 files for the object.}

\item{config_path}{path to the config file for the single-cell object. This is generated either in the config app via \code{run_config_app}, or by auto-generating a config file via \code{generate_config_yaml} and editing it by hand. For more information on using the config app, see \href{https://amc-heme.github.io/cellDIVER/articles/config_documentation.html}{here}.}

\item{enable_metadata_addition}{when TRUE, users may interactively add metadata to objects in the app. This feature is currently unstable, so it must be opted in to by setting this value to TRUE.}

\item{port}{specify a port for launching the browser. This is optional for a single deployment but required to run several instances of the browser at the same IP address. The port can be any number between 3000:8000, except for ports blocked by Google Chrome (for more information on this, see \link[shiny]{runApp}).}

\item{host}{This is passed to \code{shiny::runApp}. See the documentation of \code{host} in \link[shiny]{runApp} for more info.}

\item{launch_browser}{This is passed to \code{shiny::runApp} as \code{launch.browser}. See the documentation of \code{launch.browser} in \link[shiny]{runApp} for more info.}

\item{full_stack_trace}{when TRUE, the full stack trace of errors is logged to the console (FALSE by default).}

\item{dev_mode}{used for development and debugging. When this is TRUE, additional logging occurrs, and the status of the app is printed in the UI. This is not reccomended to be used outside of development.}
}
\description{
Initializes the main cellDIVER app.
}
\details{
For more information on setting up an cellDIVER deployment, see the \href{https://amc-heme.github.io/cellDIVER/articles/dataset_setup_walkthrough.html}{dataset setup guide} on our website.

This function does not attach its dependencies to your search path. Earlier
versions called \code{library()} on around 25 packages as a side effect of being
called; cellDIVER now declares its imports properly, so the app runs without
altering the calling environment. If you relied on, say, \code{DimPlot()} being
available at the console after launching the app, attach the package
yourself or qualify the call as \code{Seurat::DimPlot()}.
}
